move issoldout to own file
