+++ utils/getEvents.ts
@@ -125,11 +125,3 @@ export const getUpcomingBookClubEvents = async () => {
 
   return docs;
 };
-
-export const isSoldOut = (event: Event) => {
-  const ticketsSold = event.ticketsSold ?? 0;
-  const capacity = (event.location as Location)?.capacity ?? 0;
-  if (event.price === 0 || capacity === 0) return false;
-
-  return ticketsSold >= capacity;
-};
